+++ packages/x-components/src/components/index.ts
@@ -32,9 +32,7 @@ export * from './result'
 export * from './scroll'
 export { default as SlidingPanel } from './sliding-panel.vue'
 export { default as SnippetCallbacks } from './snippet-callbacks.vue'
-export { default as SortDropdown } from './sort-dropdown.vue'
-export { default as SortList } from './sort-list.vue'
-export { default as SortPickerList } from './sort-picker-list.vue'
+export * from './sort'
 
 export * from './suggestions'
 export * from './x-component.utils'
