+++ packages/x-components/src/components/sort/index.ts
@@ -1,0 +1,3 @@
+export { default as SortDropdown } from './sort-dropdown.vue'
+export { default as SortList } from './sort-list.vue'
+export { default as SortPickerList } from './sort-picker-list.vue'
